[jindo] Use configured OSS client for blob URLs
