+++ paimon-filesystems/paimon-jindo/src/main/java/org/apache/paimon/jindo/JindoFileIO.java
@@ -25,7 +25,7 @@ import org.apache.paimon.fs.HadoopOptionsProvider;
 import org.apache.paimon.fs.Path;
 import org.apache.paimon.fs.TwoPhaseOutputStream;
 import org.apache.paimon.options.Options;
-import org.apache.paimon.oss.OSSLoader;
+import org.apache.paimon.oss.OSSBlobPresigner;
 import org.apache.paimon.utils.IOUtils;
 import org.apache.paimon.utils.Pair;
 import org.apache.paimon.utils.SensitiveConfigUtils;
@@ -35,6 +35,8 @@ import com.aliyun.jindodata.common.JindoHadoopSystem;
 import com.aliyun.jindodata.dls.JindoDlsFileSystem;
 import com.aliyun.jindodata.oss.JindoOssFileSystem;
 import com.aliyun.jindodata.oss.auth.SimpleCredentialsProvider;
+import com.aliyun.oss.OSSClient;
+import com.aliyun.oss.OSSClientBuilder;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.slf4j.Logger;
@@ -67,6 +69,7 @@ public class JindoFileIO extends HadoopCompliantFileIO implements HadoopOptionsP
      */
     private static final String[] CONFIG_PREFIXES = {"fs."};
 
+    private static final String OSS_ENDPOINT = "fs.oss.endpoint";
     private static final String OSS_ACCESS_KEY_ID = "fs.oss.accessKeyId";
     private static final String OSS_ACCESS_KEY_SECRET = "fs.oss.accessKeySecret";
     private static final String OSS_SECURITY_TOKEN = "fs.oss.securityToken";
@@ -93,14 +96,12 @@ public class JindoFileIO extends HadoopCompliantFileIO implements HadoopOptionsP
     private Options hadoopOptions;
     private Options hadoopOptionsWithCache;
     private boolean allowCache = true;
-    private final FileIO blobFileIO;
+    private transient OSSClient blobClient;
 
-    public JindoFileIO() {
-        this(new OSSLoader().load(new Path("oss:///")));
-    }
+    public JindoFileIO() {}
 
-    JindoFileIO(FileIO blobFileIO) {
-        this.blobFileIO = blobFileIO;
+    JindoFileIO(OSSClient blobClient) {
+        this.blobClient = blobClient;
     }
 
     @Override
@@ -111,7 +112,6 @@ public class JindoFileIO extends HadoopCompliantFileIO implements HadoopOptionsP
     @Override
     public void configure(CatalogContext context) {
         super.configure(context);
-        blobFileIO.configure(context);
         allowCache = context.options().get(FILE_IO_ALLOW_CACHE);
         hadoopOptions = new Options();
         // read all configuration with prefix 'CONFIG_PREFIXES'
@@ -215,7 +215,30 @@ public class JindoFileIO extends HadoopCompliantFileIO implements HadoopOptionsP
     public String createBlobPresignedUrl(
             Path tableRoot, BlobDescriptor descriptor, String extension, Duration validity)
             throws IOException {
-        return blobFileIO.createBlobPresignedUrl(tableRoot, descriptor, extension, validity);
+        return OSSBlobPresigner.create(blobClient(), tableRoot, descriptor, extension, validity);
+    }
+
+    private synchronized OSSClient blobClient() {
+        if (blobClient == null) {
+            blobClient = createBlobClient(hadoopOptions);
+        }
+        return blobClient;
+    }
+
+    static OSSClient createBlobClient(Options options) {
+        String securityToken = options.get(OSS_SECURITY_TOKEN);
+        OSSClientBuilder builder = new OSSClientBuilder();
+        return (OSSClient)
+                (StringUtils.isNullOrWhitespaceOnly(securityToken)
+                        ? builder.build(
+                                options.get(OSS_ENDPOINT),
+                                options.get(OSS_ACCESS_KEY_ID),
+                                options.get(OSS_ACCESS_KEY_SECRET))
+                        : builder.build(
+                                options.get(OSS_ENDPOINT),
+                                options.get(OSS_ACCESS_KEY_ID),
+                                options.get(OSS_ACCESS_KEY_SECRET),
+                                securityToken));
     }
 
     @Override
@@ -266,7 +289,11 @@ public class JindoFileIO extends HadoopCompliantFileIO implements HadoopOptionsP
     }
 
     @Override
-    public void close() {
+    public synchronized void close() {
+        if (blobClient != null) {
+            blobClient.shutdown();
+            blobClient = null;
+        }
         if (!allowCache) {
             fsMap.values().stream().map(Pair::getKey).forEach(IOUtils::closeQuietly);
             fsMap.clear();
