+++ paimon-filesystems/paimon-jindo/src/test/java/org/apache/paimon/jindo/JindoFileIOTest.java
@@ -18,64 +18,86 @@
 
 package org.apache.paimon.jindo;
 
-import org.apache.paimon.catalog.CatalogContext;
 import org.apache.paimon.data.BlobDescriptor;
 import org.apache.paimon.fs.Path;
-import org.apache.paimon.fs.local.LocalFileIO;
 import org.apache.paimon.options.Options;
 
+import com.aliyun.oss.HttpMethod;
+import com.aliyun.oss.OSSClient;
+import com.aliyun.oss.model.ObjectMetadata;
 import org.junit.jupiter.api.Test;
 
-import java.io.IOException;
+import java.net.URI;
+import java.net.URL;
+import java.security.MessageDigest;
 import java.time.Duration;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.ArgumentMatchers.contains;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 /** Tests for {@link JindoFileIO}. */
 public class JindoFileIOTest {
 
     @Test
-    public void testCreateBlobPresignedUrlDelegates() throws Exception {
-        TestingFileIO delegate = new TestingFileIO();
-        JindoFileIO fileIO = new JindoFileIO(delegate);
-        CatalogContext context = CatalogContext.create(new Options());
-        Path tableRoot = new Path("oss://bucket/table");
-        BlobDescriptor descriptor = new BlobDescriptor("oss://bucket/table/data/file", 10, 20);
-        Duration validity = Duration.ofHours(1);
+    public void testCreateBlobClientUsesConfiguredSts() {
+        Options options = new Options();
+        options.set("fs.oss.endpoint", "https://oss.example.com");
+        options.set("fs.oss.accessKeyId", "access-key");
+        options.set("fs.oss.accessKeySecret", "access-secret");
+        options.set("fs.oss.securityToken", "security-token");
 
-        fileIO.configure(context);
-
-        assertThat(fileIO.createBlobPresignedUrl(tableRoot, descriptor, "jpg", validity))
-                .isEqualTo("https://bucket.oss.example/blob.jpg");
-        assertThat(delegate.context).isSameAs(context);
-        assertThat(delegate.tableRoot).isEqualTo(tableRoot);
-        assertThat(delegate.descriptor).isEqualTo(descriptor);
-        assertThat(delegate.extension).isEqualTo("jpg");
-        assertThat(delegate.validity).isEqualTo(validity);
+        OSSClient client = JindoFileIO.createBlobClient(options);
+        try {
+            assertThat(client.getEndpoint()).isEqualTo(URI.create("https://oss.example.com"));
+            assertThat(client.getCredentialsProvider().getCredentials().getAccessKeyId())
+                    .isEqualTo("access-key");
+            assertThat(client.getCredentialsProvider().getCredentials().getSecretAccessKey())
+                    .isEqualTo("access-secret");
+            assertThat(client.getCredentialsProvider().getCredentials().getSecurityToken())
+                    .isEqualTo("security-token");
+        } finally {
+            client.shutdown();
+        }
     }
 
-    private static class TestingFileIO extends LocalFileIO {
+    @Test
+    public void testCreateBlobPresignedUrlUsesOssClient() throws Exception {
+        OSSClient client = mock(OSSClient.class);
+        Path tableRoot = new Path("oss://bucket/table");
+        BlobDescriptor descriptor = new BlobDescriptor("oss://bucket/table/data/file", 10, 20);
+        ObjectMetadata metadata = new ObjectMetadata();
+        metadata.setContentLength(20);
+        metadata.setContentType("image/jpeg");
+        metadata.addUserMetadata(
+                "paimon-blob-descriptor-sha256", sha256Hex(descriptor.serialize()));
+        when(client.headObject(eq("bucket"), contains("_bloburl_"))).thenReturn(metadata);
+        when(client.getEndpoint()).thenReturn(URI.create("https://oss.example.com"));
+        when(client.generatePresignedUrl(eq("bucket"), anyString(), any(), eq(HttpMethod.GET)))
+                .thenAnswer(
+                        invocation ->
+                                new URL(
+                                        "https://bucket.oss.example.com/"
+                                                + invocation.getArgument(1)));
 
-        private CatalogContext context;
-        private Path tableRoot;
-        private BlobDescriptor descriptor;
-        private String extension;
-        private Duration validity;
+        JindoFileIO fileIO = new JindoFileIO(client);
+        assertThat(fileIO.createBlobPresignedUrl(tableRoot, descriptor, "jpg", Duration.ofHours(1)))
+                .startsWith("https://bucket.oss.example.com/table/data/_bloburl_");
 
-        @Override
-        public void configure(CatalogContext context) {
-            this.context = context;
+        fileIO.close();
+        verify(client).shutdown();
+    }
+
+    private static String sha256Hex(byte[] bytes) throws Exception {
+        StringBuilder result = new StringBuilder();
+        for (byte value : MessageDigest.getInstance("SHA-256").digest(bytes)) {
+            result.append(String.format("%02x", value & 0xff));
         }
-
-        @Override
-        public String createBlobPresignedUrl(
-                Path tableRoot, BlobDescriptor descriptor, String extension, Duration validity)
-                throws IOException {
-            this.tableRoot = tableRoot;
-            this.descriptor = descriptor;
-            this.extension = extension;
-            this.validity = validity;
-            return "https://bucket.oss.example/blob.jpg";
-        }
+        return result.toString();
     }
 }
